refactor: move StoreCardItem to new file
this commit also adds a convention for importing files with a `@` prefix. `types/` was created to store types
