+++ ui/hockey-teeth/src/types/store.ts
@@ -1,0 +1,8 @@
+export interface StoreItem {
+  title: string;
+  description: string;
+  image: string;
+  link: string;
+}
+
+export type StoreItemCardProps = StoreItem;
